Refactor DashResponsive to its own file
